Implement routing between item list and item detail
Use react-router to handle the routing but use in-memory routing so the
app is independent of the host's URL. Switching between list and detail
also doesn't change the URL. This means we can embed the app in
different pages with different URLs. This also should play nicely with
Tanner's implementation of iframe-based routing.