Utilize the persistent layer for individual talk
If the persistent layer is active, use that instead of the individual
player. If not, then use the individual player.

When going back to the list of talks from an individual talk, if the
audio is still playing, hand it off to the persistent player. There's a
slight kink in that transition as we only update the progress every so
often. Lowering the progress update interval can give us smoother
transition but may have performance impact. Adjust cautiously.

<PersistentPlayerProvider /> is where all the syncing and communicating
happen between the persistent player app and the consumer. It always
use the top-most window to support the iframe navigation trick.