Fix open/close state of topics accordion
Back button was not working because there were 2 states to manage the
open/close state and the wrong one was toggled.

Selecting a topic closed the accordion because the accordion was
replaced with a new instance everytime the active filter changes by the
<RenderAccordionPopular />. This component (and others alike) were
declared inside a render function so, whenever active filter changes,
a new instance is always created and rendered in place of the old one.

The pushing down of the rest of the content when the accordion is
opened is the expected behavior of a relatively-positioned DOM element.
To achieve what we need, we need to render the accordion content either
absolutely or as an element attached to a higher DOM node. While we can
try to shoe-horn this behavior to the accordion content, I think we'll
be better served by abandoning accordion and moving to popper instead.