Re-creates topic filter as a popper instead of an accordion
After fixing some critical bugs mentioned by Chris, it's evident that an
accordion is not the right tool for the job. The main limitation is the
reltaive positioning of an accordion's content.

A popper is what we need. Combined with <ClickAwayListener /> and
<Collapse /> we're able to place, position, and transition things
better.

Because it's a significantly different structure, I decide to leave the
accordion version alone in this commit and create a new component:
<TopicFilter />

Since the styling in `Filter.scss` is tightly coupled with the accordion
version, most of the styling for the new component is embedded in the
JSX. Some light styling around fonts and button states may be needed to
match it to the website theme.